rf(lanalytics): extract hooks
Move hooks for session and la handling into separated components. Add exception handling for saving.
